Final form draft 1, all working

+++ src/MovingGrumpyCat.java
@@ -1,3 +1,13 @@
+/*	Dane Jarvie
+ * 	UPI: djar004
+ * 	ID: 2521969
+ * 
+ * ======================================================================
+ *	MovingGrumpyCat.java : An extension of MovingImage.  Implements an image of
+ *	Grumpy Cat.
+ *	======================================================================
+ */
+
 import java.awt.Color;
 import java.io.File;
 import java.io.IOException;
@@ -11,7 +21,7 @@ public class MovingGrumpyCat extends MovingImage {
 		super();
 	}
 	
-	public MovingGrumpyCat(int x, int y, int pw, int w, int h,  int mw, int mh, Color bc, int pathType) {
+	public MovingGrumpyCat(int x, int y, float pw, int w, int h,  int mw, int mh, Color bc, int pathType) {
 		super(x ,y , pw, w, h ,mw ,mh ,bc, pathType);
 		try {
 			img = ImageIO.read(new File("C:\\Users\\Dane\\Documents\\UNI\\Java\\Projects\\CS230 Asst2\\src\\grumpycat2.png"));
